trunk/L2EmuProject-Game/DataPack: * L2J High Five Update * High Five protocols, thanks L2J Team * High Five packets, thanks L2J Team * High Five skills, thanks L2JDP Team * High Five skill trees, thanks L2JDP Team

+++ L2EmuProject-Game/src/main/java/net/l2emuproject/gameserver/templates/effects/EffectTemplate.java
@@ -69,7 +69,7 @@ public final class EffectTemplate
 		lambda = set.getDouble("val", 0);
 		count = Math.max(1, set.getInteger("count", 1));
 		
-		int time = set.getInteger("time", 1) * skill.getTimeMulti();
+		int time = (int) (set.getInteger("time", 1) * skill.getTimeMulti());
 		
 		if (time < 0)
 		{
